day5/requirements.txt
openai
python-dotenv
numpy
scikit-learn
tiktoken